MTM_Blog_Application/requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
dataclasses-json==0.5.8
distlib==0.3.8
Django==5.0.1
django-auto-logout==0.5.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.13.1
idna==3.6
jsons==1.6.3
marshmallow==3.20.2
marshmallow-enum==1.5.1
mypy-extensions==1.0.0
mysqlclient==2.2.4
openpyxl==3.1.2
packaging==23.2
phonepe-sdk==1.1.0
pillow==10.2.0
platformdirs==4.1.0
pycparser==2.21
pycryptodome==3.18.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing-inspect==0.9.0
typing_extensions==4.9.0
typish==1.9.3
tzdata==2023.4
tzlocal==5.2
urllib3==1.26.12
virtualenv==20.25.0
